requirements.txt
python-telegram-bot==20.7
httpx>=0.24.1
Pillow
pytesseract
python-dotenv
deep-translator>=1.11.4
requests==2.31.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.23
aiohttp==3.9.1
schedule==1.2.1
python-dateutil==2.8.2
psutil>=5.9.0 